requirements.txt
fastapi
uvicorn
python-multipart

langchain
langchain-community

openai

faiss-cpu
sentence-transformers

pypdf
python-docx

streamlit

python-dotenv
tiktoken
numpy
